% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpMetFilter.R
\name{ExpMetFilter}
\alias{ExpMetFilter}
\title{ExpMetFilter}
\usage{
ExpMetFilter(File = NULL, report = FALSE)
}
\arguments{
\item{File}{The path to, including the name and file-ending of, where the result file should be created}

\item{report}{When set to 'True' will output a PDF file with how many features were removed due to each filtering process}
}
\value{
Outputs a result file with all the peaks and sample split into whether they were filtered out or not
}
\description{
A series of filters which removes a large subset of peaks from non-targeted analysis based on user input
Specifically focused on removing exposure compounds and their metabolites from toxicological experiments
}
